The previous version was a cheat. The problem was:
mesh.V0() returns a reference to a const field so you cannot modify mesh.V0 using this function. Instead you need to use
mesh.setV0()
(see header file:
/opt/openfoam-dev/src/finiteVolume/fvMesh/fvMesh.H
)

So your line:
mesh.V0() == mesh.V0()*A.oldTime()
wasn't doing anything. Instead you were incorporating A.oldTime into T BUT YOU WEREN'T INCORPORATING IT INTO uniT. If you want a method that preserves uniform fields and has conservation then, in this code, you need to treat uniT in the same way as T

Instead I did:
    if (resetA)
    {
        mesh.setV0() = A.oldTime()*mesh.V0();
        A.oldTime() == dimensionedScalar("", dimless, scalar(1));
    }
and then the rest of the code as before. This is actually identical to the version that doesn't reset A. So you get conservation of AT and you get the uniT property but A diverges as you are solving exactly the same equation as before for A. A.oldTime is just incorporated into Vold. So I am afraid we are back to where we were before. If you did:

T.oldTime() == T.oldTime()*A.oldTime();
and
uniT.oldTime() == uniT.oldTime()*A.oldTime();

then you would no longer get the uniT property.

I have pushed a version which just resets A and loses conservation.